Rename config/cli submodule to config-cli
