+++ packages/core/config-cli.d.ts
@@ -1,0 +1,1 @@
+﻿export * from './types/config-cli/index';
