+++ packages/core/config-cli.js
@@ -1,0 +1,1 @@
+﻿module.exports = require('./dist/cjs/config-cli/index');
