--- packages/core/config/cli.d.ts
@@ -1,1 +1,0 @@
-﻿export * from '../types/config/cli/index';
