+++ packages/core/tsconfig.json
@@ -24,7 +24,7 @@
     "tools.d.ts",
     "personalize.d.ts",
     "feaas.d.ts",
-    "config/cli.d.ts",
+    "config-cli.d.ts",
     "src/**/*.test.ts",
     "src/test-data/*",
     "src/tools/templating/test-data/*",
