--- packages/nextjs/config/cli.js
@@ -1,1 +1,0 @@
-﻿module.exports = require('../dist/cjs/config/cli/index');
