+++ packages/nextjs/tsconfig.json
@@ -25,7 +25,7 @@
     "client.d.ts",
     "utils.d.ts",
     "tools.d.ts",
-    "config/cli.d.ts",
+    "config-cli.d.ts",
     "src/tests/*",
     "src/test-data/*",
     "**/*.test.ts",
